+++ src/Http/Controllers/PaymentGateways/PaymentGateway.php
@@ -88,16 +88,17 @@ abstract class PaymentGateway extends WebController
     /**
      * Create the order in our database.
      */
-    protected function createOrder(string $id, Collection $items, string $orderNumber, Customer $customer, string $totalPrice, ?string $method = null): Order
+    protected function createOrder(string $id, Collection $items, string $orderNumber, Customer $customer, string $totalPrice, ?string $method = null, ?Collection $shippings = null): Order
     {
         $order = Order::create([
             'id'           => $id,
+            'method'       => $method,
             'status'       => 'created',
             'customer'     => $customer,
             'total_amount' => $totalPrice,
-            'method'       => $method,
             'number'       => $orderNumber,
             'items'        => (new ItemCollection($items))->items,
+            'shippings'    => $shippings ? json_encode($shippings->toArray()) : collect(), // TODO: Require shippings with the next major release.
         ]);
 
         event(new OrderCreated($order));