+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## Upcoming Release
+- Shippings will be saved in the order. (#149)
+
 ## 3.4.1 (16. Jan 2021)
 - Respect saved listing preferences on the orders overview. (#181)
 - Add the missing translation attribute 'items count'.